Organised valueTypes into separate files
